Grouping WC external wallet signing functions
